refactor(api): move entity types definition to domain
