+++ apps/api/src/domain/repositories/interfaces/repository-provider.ts
@@ -1,11 +1,11 @@
-import { EntityName } from 'apps/api/src/persistence/database/get-arango-collection-ids';
 import { InstanceFactory } from 'apps/api/src/persistence/repositories/repository-for-entity';
 import { Entity } from '../../models/entity';
+import { EntityType } from '../../types/entityType';
 import { IRepositoryForEntity } from './repository-for-entity';
 
 export interface IRepositoryProvider {
   forEntity: <TEntity extends Entity>(
-    entityName: EntityName,
+    entityType: EntityType,
     instanceFactory: InstanceFactory<TEntity>
   ) => IRepositoryForEntity<TEntity>;
 }
