+++ apps/api/src/lib/types/valueType.ts
@@ -1,0 +1,1 @@
+export type ValueType<T extends object> = T[keyof T];
