+++ apps/api/src/persistence/database/types/ArangoCollectionId.ts
@@ -1,0 +1,3 @@
+const arangoCollectionIds = ['terms', 'vocabulary_lists'] as const;
+
+export type ArangoCollectionID = typeof arangoCollectionIds[number];
