+++ apps/api/src/persistence/repositories/repository.provider.ts
@@ -1,11 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { Entity } from '../../domain/models/entity';
 import { IRepositoryProvider } from '../../domain/repositories/interfaces/repository-provider';
+import { EntityType } from '../../domain/types/entityType';
 import { DatabaseProvider } from '../database/database.provider';
-import {
-  EntityName,
-  getArangoCollectionID,
-} from '../database/get-arango-collection-ids';
+import { getArangoCollectionID } from '../database/get-arango-collection-ids';
 import { InstanceFactory, RepositoryForEntity } from './repository-for-entity';
 
 @Injectable()
@@ -13,12 +11,12 @@ export class RepositoryProvider implements IRepositoryProvider {
   constructor(private databaseProvider: DatabaseProvider) {}
 
   forEntity<TEntity extends Entity>(
-    entityName: EntityName,
+    entityType: EntityType,
     instanceFactory: InstanceFactory<TEntity>
   ) {
     return new RepositoryForEntity<TEntity>(
       this.databaseProvider,
-      getArangoCollectionID(entityName),
+      getArangoCollectionID(entityType),
       instanceFactory
     );
   }
